Reintegração TipoCausa,TipoTarefa,TipoDespesa.Gravar Desp,Hon.
Foram refeitas as classes TipoCausa, TipoTarefa e TipoDespesa.
Desenvolvido o Gravar de LançamentoDepesa e LançamentoHonorário com
testes.

Mudança de alguns tipos de atributos.

+++ NAC_INTEGRADA_ALFAAD/src/br/com/fiap/beans/Pessoa.java
@@ -4,12 +4,12 @@ import java.util.List;
 
 public abstract class Pessoa {
 
-	private int codigo;
+	private long codigo;
 	private String nome;
 	private List <Endereco> endereco;
 	private List <Telefone> telefone;
 	
-	public Pessoa(int codigo, String nome, List<Endereco> endereco, List<Telefone> telefone) {
+	public Pessoa(long codigo, String nome, List<Endereco> endereco, List<Telefone> telefone) {
 		super();
 		this.codigo = codigo;
 		this.nome = nome;
@@ -19,11 +19,11 @@ public abstract class Pessoa {
 
 	public Pessoa() {}
 
-	public int getCodigo() {
+	public long getCodigo() {
 		return codigo;
 	}
 
-	public void setCodigo(int codigo) {
+	public void setCodigo(long codigo) {
 		this.codigo = codigo;
 	}
